+++ STM32DeviceExample/STM32DeviceExample/ViewController.swift
@@ -80,7 +80,7 @@ class ViewController: NSViewController, NSComboBoxDataSource {
         return self.devices[index].deviceInfo.name
     }
 
-    func usbConnected(notification: NSNotification) {
+    @objc func usbConnected(notification: NSNotification) {
         guard let nobj = notification.object as? NSDictionary else {
             return
         }
@@ -95,7 +95,7 @@ class ViewController: NSViewController, NSComboBoxDataSource {
         }
     }
     
-    func usbDisconnected(notification: NSNotification) {
+    @objc func usbDisconnected(notification: NSNotification) {
         guard let nobj = notification.object as? NSDictionary else {
             return
         }