+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.3
+
+- Updated all projects to Switf 4
+- Created Wiki documentation [here](https://github.com/Arti3DPlayer/USBDeviceSwift/wiki)
+- Added Serial device support
+
 ## 1.0.2
 - Explicitly set Swift version project-wide
 - Explicitly set Swift version in the pod spec (this was most likely what fixed #3)